refactored code and added some testcases
