+++ src/utils/addStatement/checkBlankString.js
@@ -1,0 +1,5 @@
+function isBlankString(str) {
+    return !str || /^\s*$/.test(str);
+}
+
+module.exports = isBlankString
